Updated Sample Data Plugin
The first edition overwrote the current admin user so after importing the data you were logged out and no way of getting back in. This adds a first step which looks for the admin user, moves the id to 5 and then imports the data. You are then forced to re-login but it does now work.