Complete Sync
This is a complete sync of my development environment including stage parts of User Control Panel, Submit Extensions, Submit Ticket and Ticket Editing. I am uploading it so I can break it into smaller tasks for others to help with. This also includes a version of the JTemplate with the menu currently hardcoded for testing.